+++ packages/@expo/config-types/src/ExpoConfig.ts
@@ -71,7 +71,7 @@ export interface ExpoConfig {
    */
   icon?: string;
   /**
-   * Configuration for remote (push) notifications.
+   * @deprecated in favor of expo-notifications config plugin. Configuration for remote (push) notifications.
    */
   notification?: {
     /**