+++ packages/expo-ui/CHANGELOG.md
@@ -10,7 +10,16 @@
 
 ### 💡 Others
 
+## 0.2.0-alpha.9 — 2025-09-03
+
+_This version does not introduce any user-facing changes._
+
+## 0.2.0-alpha.8 — 2025-09-02
+
+### 💡 Others
+
 - Used convertibles to process modifiers' parameters. ([#39231](https://github.com/expo/expo/pull/39231) by [@kudo](https://github.com/kudo))
+- Migrated `AnimationModifier` to convertibles. ([#39326](https://github.com/expo/expo/pull/39326) by [@kudo](https://github.com/kudo))
 
 ## 0.2.0-alpha.7 — 2025-08-31
 
@@ -25,6 +34,7 @@ _This version does not introduce any user-facing changes._
 ### 🎉 New features
 
 - [iOS] Add `animation` modifier. ([#38954](https://github.com/expo/expo/pull/38954) by [@nishan](https://github.com/intergalacticspacehighway))
+- [iOS] Add `GlassEffectContainer`, `glassEffectId` and `Namespaces` support. ([#39070](https://github.com/expo/expo/pull/39070) by [@nishan](https://github.com/intergalacticspacehighway))
 
 ### 🐛 Bug fixes